Update on "[dynamo] Log all compilation metrics in torch._dynamo.metrics"
Motivation: for testing in dynamo, it would be nice to have an equivalent to torch._inductor.metrics. We already have log_compilation_event, but that only prints logs or dumps to a database (in fbcode); these aren't appropriate for unit tests.

This change: adds torch._dynamo.metrics, and then changes convert_frame so that it calls torch._dynamo.metrics.record_compilation_metrics instead. To prevent a super large list from forming, use a deque with a max length; then if the max length is exceeded, the oldest records will be dropped.

cc voznesenskym penguinwu EikanWang jgong5 Guobing-Chen XiaobingSuper zhuhaozhe blzheng wenzhe-nrv jiayisunx chenyang78 aakhundov kadeng

[ghstack-poisoned]
